ChipQA/ChipQA_tmp/requirements.txt
joblib 
numpy
cv2
queue
scipy.ndimage
sys
matplotlib
numba 
argparse
